Print message for used supervisors
Supervisors have to implement a method `getName()` which returns a name
of the supervisor. This name is used to print a message when the
supervisor is added to the chain.

+++ path_follower/src/supervisor/pathlookout.cpp
@@ -248,7 +248,7 @@ vector<Obstacle> PathLookout::lookForObstaclesInScans()
     observed_obstacles.reserve(obstacle_points.size());
     for (vector<vector<cv::Point2f> >::const_iterator it = obstacle_points.begin(); it != obstacle_points.end(); ++it) {
         // ignore clusters, that are too small
-        if (it->size() < opt_.min_number_of_points()) {
+        if (it->size() < (size_t)opt_.min_number_of_points()) {
             continue;
         }
